9-HW-DS561/requirements.txt
Flask==3.0.0
gunicorn
google-cloud-storage
google-cloud-logging
google-cloud-pubsub